requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
pywin32
Pillow
